New 5.x build with many community enhancements and .NET 5.0 targeting (#42)
* New 5.x build with many community enhancements and .NET 5.0 targeting.  NOTE!  DROPING .NET Framework Support in this major release.

* Update README.md

* Allow the ClamClient to support connecting to a ClamAV server based on IP Address. This use case can happen if the ClamAV server is only accessible via a private IP address. The TcpClient.ConnectAsync(hostname assumes hostname is a DNS address.

Updated the readme and example programs to also use the IPAddress version of the ClamClient.

* Allow the ClamClient to support connecting to a ClamAV server based on IP Address. This use case can happen if the ClamAV server is only accessible via a private IP address. The TcpClient.ConnectAsync(hostname assumes hostname is a DNS address.

Updated the readme and example programs to also use the IPAddress version of the ClamClient.

* Prevent check of stream position on streams that don't support seeking, like http body content...
