+++ README.md
@@ -10,13 +10,14 @@ ClamAV Server, also known as clamd.  It is a free, open-source virus scanner.  W
 
 ## Directions
 1. Add the nuget package to your project.
-2. Create a nClam.ClamClient object, passing it the hostname and port of the ClamAV server.
+2. Create a nClam.ClamClient object, passing it the hostname (or IP address) and port of the ClamAV server.
 3. Scan!
 
 # Code Example
 ```csharp
 using System;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using nClam;
 
@@ -25,6 +26,7 @@ class Program
     static async Task Main(string[] args)
     {
         var clam = new ClamClient("localhost", 3310);
+		// or var clam = new ClamClient(IPAddress.Parse("127.0.0.1"), 3310);
         var scanResult = await clam.ScanFileOnServerAsync("C:\\test.txt");  //any file you would like!
 
         switch (scanResult.Result)
@@ -49,7 +51,7 @@ class Program
 For directions on setting up ClamAV as a Windows Service, check out [this blog post](http://architectryan.com/2011/05/19/nclam-a-dotnet-library-to-virus-scan/).
 
 # Test Application
-For more information about how to use nClam, you can look at the nClam.ConsoleTest project's [Program.cs](https://github.com/tekmaven/nClam/blob/master/nClam.ConsoleTestC2/Program.cs).
+For more information about how to use nClam, you can look at the nClam.ConsoleTest project's [Program.cs](https://github.com/tekmaven/nClam/blob/master/nClam.ConsoleTest/Program.cs).
 
 # Contributing
 I accept PRs!  We have had several contributors help maintain this library by fixing bugs, introducing async support, and moving to .NET Core.  Thank you to all the contributors!
