+++ nClam.Tests/ClamScanResultTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Net;
 using Xunit;
 
 namespace nClam.Tests
@@ -91,6 +92,15 @@ namespace nClam.Tests
 			var client = new ClamClient("localhost");
 			var result = client.SendAndScanFileAsync(new MemoryStream(System.Text.Encoding.Default.GetBytes(Eicartestcase)));
 			Assert.Equal(ClamScanResults.VirusDetected,result.Result.Result);
-		}
+        }
+
+        [Fact(Skip = "Requires ClamAV running on 127.0.0.1:3310 ")]
+        public void TestSendIPAsyncTest()
+        {
+            string Eicartestcase = @"X5O!P%@AP[4\PZX54(P^)7CC)7}$EICAR-STANDARD-ANTIVIRUS-TEST-FILE!$H+H*";
+            var client = new ClamClient(IPAddress.Parse("127.0.0.1"));
+            var result = client.SendAndScanFileAsync(new MemoryStream(System.Text.Encoding.Default.GetBytes(Eicartestcase)));
+            Assert.Equal(ClamScanResults.VirusDetected, result.Result.Result);
+        }
     }
 }
